+++ pkgs/intl4x/lib/src/list_format/list_format_stub_4x.dart
@@ -1,0 +1,12 @@
+// Copyright (c) 2023, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import '../../list_format.dart';
+import '../data.dart';
+import '../locale/locale.dart';
+import 'list_format_impl.dart';
+
+ListFormatImpl getListFormatter4X(
+        Locale locale, Data data, ListFormatOptions options) =>
+    throw UnimplementedError('Cannot use ICU4X in web environments.');